requirements.txt
django==1.11
djangorestframework
markdown
django-filter
psycopg2
